Added casts to avoid SDCC warnings on implicit type conversion

+++ Tek/Firmware/Tek_PIC16F688_SDCC/ZX81_Keycodes.h
@@ -20,53 +20,53 @@
 
 // Mapas das teclas
 //    Tecla   coluna  linha       
-#define _KEY_1	  (0<<3) + 0 
-#define _KEY_2	  (1<<3) + 0 
-#define _KEY_3	  (2<<3) + 0 
-#define _KEY_4	  (3<<3) + 0 
-#define _KEY_5	  (4<<3) + 0 
-#define _KEY_6	  (4<<3) + 3 
-#define _KEY_7	  (3<<3) + 3 
-#define _KEY_8	  (2<<3) + 3 
-#define _KEY_9	  (1<<3) + 3 
-#define _KEY_0	  (0<<3) + 3 
-#define _KEY_Q	  (0<<3) + 1 
-#define _KEY_W	  (1<<3) + 1 
-#define _KEY_E	  (2<<3) + 1 
-#define _KEY_R	  (3<<3) + 1 
-#define _KEY_T	  (4<<3) + 1 
-#define _KEY_Y	  (4<<3) + 4 
-#define _KEY_U	  (3<<3) + 4 
-#define _KEY_I	  (2<<3) + 4 
-#define _KEY_O	  (1<<3) + 4 
-#define _KEY_P	  (0<<3) + 4 
-#define _KEY_A	  (0<<3) + 2 
-#define _KEY_S	  (1<<3) + 2 
-#define _KEY_D	  (2<<3) + 2 
-#define _KEY_F	  (3<<3) + 2 
-#define _KEY_G	  (4<<3) + 2 
-#define _KEY_H	  (4<<3) + 6 
-#define _KEY_J	  (3<<3) + 6 
-#define _KEY_K	  (2<<3) + 6 
-#define _KEY_L	  (1<<3) + 6 
-#define _KEY_NL   (0<<3) + 6 // New Line
-#define _KEY_SH   (0<<3) + 5 // Shift
-#define _KEY_Z	  (1<<3) + 5 	
-#define _KEY_X	  (2<<3) + 5 	
-#define _KEY_C	  (3<<3) + 5 	
-#define _KEY_V	  (4<<3) + 5 	
-#define _KEY_B	  (4<<3) + 7 
-#define _KEY_N	  (3<<3) + 7 
-#define _KEY_M	  (2<<3) + 7 
-#define _KEY_DOT  (1<<3) + 7 
-#define _KEY_SP	  (0<<3) + 7 
-#define _KEY_NONE (7<<3) + 7   
+#define _KEY_1	  (uint8_t) ((0<<3) + 0) 
+#define _KEY_2	  (uint8_t) ((1<<3) + 0) 
+#define _KEY_3	  (uint8_t) ((2<<3) + 0) 
+#define _KEY_4	  (uint8_t) ((3<<3) + 0) 
+#define _KEY_5	  (uint8_t) ((4<<3) + 0) 
+#define _KEY_6	  (uint8_t) ((4<<3) + 3) 
+#define _KEY_7	  (uint8_t) ((3<<3) + 3) 
+#define _KEY_8	  (uint8_t) ((2<<3) + 3) 
+#define _KEY_9	  (uint8_t) ((1<<3) + 3) 
+#define _KEY_0	  (uint8_t) ((0<<3) + 3) 
+#define _KEY_Q	  (uint8_t) ((0<<3) + 1) 
+#define _KEY_W	  (uint8_t) ((1<<3) + 1) 
+#define _KEY_E	  (uint8_t) ((2<<3) + 1) 
+#define _KEY_R	  (uint8_t) ((3<<3) + 1) 
+#define _KEY_T	  (uint8_t) ((4<<3) + 1) 
+#define _KEY_Y	  (uint8_t) ((4<<3) + 4) 
+#define _KEY_U	  (uint8_t) ((3<<3) + 4) 
+#define _KEY_I	  (uint8_t) ((2<<3) + 4) 
+#define _KEY_O	  (uint8_t) ((1<<3) + 4) 
+#define _KEY_P	  (uint8_t) ((0<<3) + 4) 
+#define _KEY_A	  (uint8_t) ((0<<3) + 2) 
+#define _KEY_S	  (uint8_t) ((1<<3) + 2) 
+#define _KEY_D	  (uint8_t) ((2<<3) + 2) 
+#define _KEY_F	  (uint8_t) ((3<<3) + 2) 
+#define _KEY_G	  (uint8_t) ((4<<3) + 2) 
+#define _KEY_H	  (uint8_t) ((4<<3) + 6) 
+#define _KEY_J	  (uint8_t) ((3<<3) + 6) 
+#define _KEY_K	  (uint8_t) ((2<<3) + 6) 
+#define _KEY_L	  (uint8_t) ((1<<3) + 6) 
+#define _KEY_NL   (uint8_t) ((0<<3) + 6) // New Line
+#define _KEY_SH   (uint8_t) ((0<<3) + 5) // Shift
+#define _KEY_Z	  (uint8_t) ((1<<3) + 5) 	
+#define _KEY_X	  (uint8_t) ((2<<3) + 5) 	
+#define _KEY_C	  (uint8_t) ((3<<3) + 5) 	
+#define _KEY_V	  (uint8_t) ((4<<3) + 5) 	
+#define _KEY_B	  (uint8_t) ((4<<3) + 7) 
+#define _KEY_N	  (uint8_t) ((3<<3) + 7) 
+#define _KEY_M	  (uint8_t) ((2<<3) + 7) 
+#define _KEY_DOT  (uint8_t) ((1<<3) + 7) 
+#define _KEY_SP	  (uint8_t) ((0<<3) + 7) 
+#define _KEY_NONE (uint8_t) ((7<<3) + 7)   
 
 // Flag da tecla de SHIFT
-#define _SHIFT 0x80
+#define _SHIFT  (uint8_t)0x80
 
-#define _SH_COL  0 
-#define _SH_LINE 5
+#define _SH_COL   (uint8_t)0 
+#define _SH_LINE  (uint8_t)5
 
-#define _EX_COL  1 
-#define _EX_LINE 7
+#define _EX_COL   (uint8_t)1 
+#define _EX_LINE  (uint8_t)7
